Update check-mathml.js
We should probably come up with a global object name - like THREE does - so we don't trounce other libraries.

	script.type = "text/javascript";

so old school. Is there a browser that supports WebGL that needs this, I don;t think so.

//	document.getElementsByTagName("head")[0].appendChild(script);
	document.body.appendChild( scriptMJ );	

Mr.doob lods his scripts to body. Why can't we?

Could this file be in a folder with a read me that explains is whys and wherefores?